tout

+++ src/core/ConcreteCardVisitor.java
@@ -15,20 +15,20 @@ public class ConcreteCardVisitor implements CardVisitor {
 		} else {
 			if (vlibreCard.getTimeCredit() == 0) {
 				if(bike instanceof MechanicalBike) {
-					mechanicalCost = (rideDuration - 60) / 60;
+					mechanicalCost = (double) (rideDuration - 60) / 60;
 					return mechanicalCost;}
 				else {
-					electricalCost = (rideDuration - 60) / 30;
+					electricalCost = (double) (rideDuration - 60) / 30;
 					return electricalCost;}
 			} else {
 				excessTimeBalance = 60 - (rideDuration - vlibreCard.getTimeCredit());
 				if (excessTimeBalance <= 0) {
 					vlibreCard.setTimeCredit(0);
 					if(bike instanceof MechanicalBike) {
-						mechanicalCost = -excessTimeBalance / 60;
+						mechanicalCost = (double) -excessTimeBalance / 60;
 						return mechanicalCost;}
 					else {
-						electricalCost = -excessTimeBalance / 30;
+						electricalCost = (double) -excessTimeBalance / 30;
 						return electricalCost;}
 				} else {
 					vlibreCard.setTimeCredit((int) excessTimeBalance);
@@ -52,13 +52,13 @@ public class ConcreteCardVisitor implements CardVisitor {
 		}
 		else {
 			if(vmaxCard.getTimeCredit() == 0) {
-				rideCost = (rideDuration - 60)/60;
+				rideCost = (double) (rideDuration - 60)/60;
 				return rideCost;
 			}
 			else {
 				excessTimeBalance = 60 - (rideDuration - vmaxCard.getTimeCredit());
 				if(excessTimeBalance <= 0) {
-					rideCost = -excessTimeBalance/60;
+					rideCost = (double) -excessTimeBalance/60;
 					vmaxCard.setTimeCredit(0);
 					return rideCost;
 				}
